chore: fix typo, move types to types.py
